Simplify types hierarchy
